refactor: 계정과 권한 결합 분리 카프카 또는 WebClient로 연동할 예정 앞으로 엔티티 간 모든 연관관계는 끊을 예정입니다.

+++ user/src/main/java/com/kalgooksoo/user/service/DefaultUserService.java
@@ -1,8 +1,10 @@
 package com.kalgooksoo.user.service;
 
 import com.kalgooksoo.user.command.UpdateUserCommand;
+import com.kalgooksoo.user.domain.Authority;
 import com.kalgooksoo.user.domain.User;
 import com.kalgooksoo.user.exception.UsernameAlreadyExistsException;
+import com.kalgooksoo.user.repository.AuthorityRepository;
 import com.kalgooksoo.user.repository.UserRepository;
 import com.kalgooksoo.user.search.UserSearch;
 import com.kalgooksoo.user.value.ContactNumber;
@@ -27,19 +29,43 @@ import static com.kalgooksoo.user.specification.UserSpecification.*;
 public class DefaultUserService implements UserService {
 
     private final UserRepository userRepository;
+
+    private final AuthorityRepository authorityRepository;
+
     private final PasswordEncoder passwordEncoder;
 
     @Override
-    public User create(User user) throws UsernameAlreadyExistsException {
+    public User createUser(User user) throws UsernameAlreadyExistsException {
         if (userRepository.exists(usernameEquals(user.getUsername()))) {
             throw new UsernameAlreadyExistsException(user.getUsername(), "계정이 이미 존재합니다");
         }
         user.changePassword(passwordEncoder.encode(user.getPassword()));
+        User savedUser = userRepository.save(user);
+        authorityRepository.save(Authority.create(savedUser.getId(), "ROLE_USER"));
         return userRepository.save(user);
     }
 
     @Override
+    public User createAdmin(User user) throws UsernameAlreadyExistsException {
+        if (userRepository.exists(usernameEquals(user.getUsername()))) {
+            throw new UsernameAlreadyExistsException(user.getUsername(), "계정이 이미 존재합니다");
+        }
+        user.changePassword(passwordEncoder.encode(user.getPassword()));
+        User savedUser = userRepository.save(user);
+        authorityRepository.save(Authority.create(savedUser.getId(), "ROLE_ADMIN"));
+        return savedUser;
+    }
+
+    /**
+     * 계정 수정
+     *
+     * @param id      계정 식별자
+     * @param command 수정 명령
+     * @return 수정된 계정
+     */
+    @Override
     public User update(String id, UpdateUserCommand command) {
+        // FIXME UpdateUserCommand는 표현 계층에 의존적이라 도메인 계층에 노출되어서는 안될 것 같습니다.
         Optional<User> foundUser = userRepository.findById(id);
         if (foundUser.isEmpty()) {
             throw new NoSuchElementException("계정을 찾을 수 없습니다.");
@@ -83,8 +109,16 @@ public class DefaultUserService implements UserService {
         return userRepository.findAll(specification, pageable);
     }
 
+    /**
+     * 계정 삭제
+     * 삭제 연산은 권한을 가진자에게만 허용할 예정이라 노출해도 무방할 것 같다고 판단하여 예외 처리를 하였습니다.
+     *
+     * @param id 계정 식별자
+     */
     @Override
     public void deleteById(String id) {
+        if (!userRepository.exists(idEquals(id)))
+            throw new NoSuchElementException("계정을 찾을 수 없습니다.");
         userRepository.deleteById(id);
     }
 
@@ -95,7 +129,7 @@ public class DefaultUserService implements UserService {
             User user = foundUser.get();
             user.changePassword(passwordEncoder.encode(password));
         } else {
-            throw new NoSuchElementException("User not found");
+            throw new NoSuchElementException("계정을 찾을 수 없습니다.");
         }
     }
